Create wmape_group_example.py
For some reason no one has posted code online on WMAPE in pandas, nor on how to use WMAPE on different groups. To make life easier for everyone, I have not only posted some code, but a Stack Overflow link at the bottom of the python script to do WMAPE on multiple forecasts on one actual in the same dataframe.